# zip-repl/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GEF Player Statistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List all matches
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a new match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchRequest"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
    put:
      operationId: updateMatch
      tags: [matches]
      summary: Update match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchRequest"
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
    delete:
      operationId: deleteMatch
      tags: [matches]
      summary: Delete match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /h2h/player:
    get:
      operationId: getPlayerH2H
      tags: [matches]
      summary: Get H2H for two players
      parameters:
        - name: player1Id
          in: query
          required: true
          schema:
            type: integer
        - name: player2Id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: H2H stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerH2HStats"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        teamId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        marketValue:
          type: number
          nullable: true
        salary:
          type: number
          nullable: true
        matchesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        goalsScored:
          type: integer
        goalsConceded:
          type: integer
        mvpCount:
          type: integer
        overallRating:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - matchesPlayed
        - wins
        - losses
        - draws
        - goalsScored
        - goalsConceded
        - mvpCount
        - overallRating
        - createdAt
...
    Match:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        team1Id:
          type: integer
        team2Id:
          type: integer
        team1Name:
          type: string
        team2Name:
          type: string
        team1Score:
          type: integer
        team2Score:
          type: integer
        playerMatchups:
          type: array
          items:
            $ref: "#/components/schemas/PlayerMatchup"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - date
        - team1Id
        - team2Id
        - team1Name
        - team2Name
        - team1Score
        - team2Score
        - playerMatchups
        - createdAt
    PlayerMatchup:
      type: object
      properties:
        id:
          type: integer
        matchId:
          type: integer
        player1Id:
          type: integer
        player2Id:
          type: integer
        player1Name:
          type: string
        player2Name:
          type: string
        player1Goals:
          type: integer
        player2Goals:
          type: integer
        mvpPlayerId:
          type: integer
          nullable: true
      required:
        - id
        - matchId
        - player1Id
        - player2Id
        - player1Name
        - player2Name
        - player1Goals
        - player2Goals
    CreateMatchRequest:
      type: object
      properties:
        date:
          type: string
        team1Id:
          type: integer
        team2Id:
          type: integer
        team1Score:
          type: integer
        team2Score:
          type: integer
        playerMatchups:
          type: array
          items:
            $ref: "#/components/schemas/CreatePlayerMatchupRequest"
        notes:
          type: string
          nullable: true
      required:
        - date
        - team1Id
        - team2Id
        - team1Score
        - team2Score
        - playerMatchups
    CreatePlayerMatchupRequest:
      type: object
      properties:
        player1Id:
          type: integer
        player2Id:
          type: integer
        player1Goals:
          type: integer
        player2Goals:
          type: integer
        mvpPlayerId:
          type: integer
          nullable: true
      required:
        - player1Id
        - player2Id
        - player1Goals
        - player2Goals
    UpdateMatchRequest:
      type: object
      properties:
        date:
          type: string
        team1Score:
          type: integer
        team2Score:
          type: integer
        playerMatchups:
          type: array
          items:
            $ref: "#/components/schemas/CreatePlayerMatchupRequest"
        notes:
          type: string
          nullable: true
    H2HRecord:
      type: object
      properties:
        player1Id:
          type: integer
        player2Id:
          type: integer
        player1Name:
          type: string
        player2Name:
          type: string
        player1ImageUrl:
          type: string
          nullable: true
        player2ImageUrl:
          type: string
          nullable: true
        totalGames:
          type: integer
        player1Wins:
          type: integer
        player2Wins:
          type: integer
        draws:
          type: integer
        player1Goals:
          type: integer
        player2Goals:
          type: integer
        rivalryScore:
          type: number
      required:
        - player1Id
        - player2Id
        - player1Name
        - player2Name
        - totalGames
        - player1Wins
        - player2Wins
        - draws
        - player1Goals
        - player2Goals
        - rivalryScore
    PlayerH2HStats:
      type: object
      properties:
        player1:
          $ref: "#/components/schemas/Player"
        player2:
          $ref: "#/components/schemas/Player"
        h2h:
          $ref: "#/components/schemas/H2HRecord"
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
      required:
        - player1
        - player2
        - h2h
        - matches
...